style: :truck: renamed chat controller file to camel case
